refactor(data-view): move logic from data-view-table.component to a data.facade and use it in data-view-chart
note: this way the same data is shared between the two, but currently 2 dropdown instances are created where the selected value is not updated
