+++ libs/feature/record/src/lib/data-view-chart/data-view-chart.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core'
 import { ChartType } from 'chart.js'
 import { BehaviorSubject } from 'rxjs'
-import { DATA_MOCK } from './data.mock'
+import { DataFacade } from '../state/data.facade'
 
 @Component({
   selector: 'gn-ui-data-view-chart',
@@ -10,9 +10,14 @@ import { DATA_MOCK } from './data.mock'
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DataViewChartComponent {
-  chartData$ = new BehaviorSubject(DATA_MOCK)
   xAxis$ = new BehaviorSubject('id')
   yAxis$ = new BehaviorSubject('pop')
   label$ = new BehaviorSubject('pop par id')
   chartType$ = new BehaviorSubject<ChartType>('bar')
+
+  constructor(protected dataFacade: DataFacade) {}
+
+  selectLinkToDisplay(link: number) {
+    this.dataFacade.selectedLinkIndex$.next(link)
+  }
 }
