--- libs/feature/record/src/lib/data-view-chart/data.mock.ts
@@ -1,737 +1,0 @@
-export const DATA_MOCK = [
-  {
-    id: "Communauté d'agglomération",
-    idurba: '200067965_SCOT_00000000',
-    nom: 'SCOT DE LA REGION DE COMPIEGNE ET DE LA BASSE AUTOMNE',
-    etiquette: 'REGION DE COMPIEGNE ET DE LA BASSE AUTOMNE',
-    etiquetmin: 'Région de Compiègne et de la Basse Automne',
-    siren_nom:
-      'CA AGGLOMERATION DE LA REGION DE COMPIEGNE ET DE LA BASSE AUTOMNE',
-    siren: '200067965',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en élaboration',
-    etat_proc: "Délibération d'élaboration du SCOT",
-    typeproc: '01',
-    typeproc_t: 'Elaboration',
-    typeproc_d: '2018-11-15',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'https://www.agglo-compiegne.fr/scot',
-    siteweb2: '',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: '',
-    nbre_com: '22',
-    pop: '84703',
-    surf_km2: '265',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Pôle métropolitain',
-    idurba: '200082063_SCOT_00000000',
-    nom: 'SCOT DU GRAND AMIENOIS',
-    etiquette: 'GRAND AMIENOIS',
-    etiquetmin: 'Grand Amiénois',
-    siren_nom: 'POLE METROPOLITAIN DU GRAND AMIENOIS',
-    siren: '200082063',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Délibération de révision du SCOT',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2018-12-19',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'http://www.aduga.org/index.php?lvl=cmspage&pageid=6&id_rubrique=112',
-    siteweb2: '',
-    dialogue_n: '9',
-    dialogue_t: 'Grand Amiénois - Grand Roye',
-    sdci: 'oui',
-    nbre_com: '466',
-    pop: '384315',
-    surf_km2: '3621',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '200009918_SCOT_00000000',
-    nom: 'SCOT BASSIN DU CREILLOIS ET VALLEES BRETHOISE',
-    etiquette: 'BASSIN CREILLOIS ET VALLEES BRETHOISE',
-    etiquetmin: 'Bassin Creillois et vallées Brethoises',
-    siren_nom: 'SYNDICAT MIXTE DU BASSIN CREILLOIS ET DES VALLEES BRETHOISE',
-    siren: '200009918',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Diagnostic',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2017-07-04',
-    diag_d: '2020-01-01',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'https://www.smbcvb.fr/',
-    siteweb2: '',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: '',
-    nbre_com: '21',
-    pop: '111208',
-    surf_km2: '132',
-    info_compl: '',
-    modif_date: '2021-09-02',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Communauté de communes',
-    idurba: '246000921_SCOT_00000000',
-    nom: "SCOT DU PAYS D'OISE ET D'HALATTE",
-    etiquette: "PAYS D'OISE ET D'HALATTE",
-    etiquetmin: "Pays d'Oise et d'Halatte",
-    siren_nom: 'COMMUNAUTE DE COMMUNES DES PAYS OISE ET HALATTE',
-    siren: '246000921',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: "Projet d'Aménagement et de Développement Durable",
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2018-04-24',
-    diag_d: '2019-07-01',
-    padd_d: '2020-12-01',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'https://www.ccpoh.fr/amenager-le-territoire/planification/scot/revision-du-scot-en-cours',
-    siteweb2: '',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: '',
-    nbre_com: '17',
-    pop: '34798',
-    surf_km2: '140',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '256204082_SCOT_00000000',
-    nom: 'SCOT LENS-LIEVIN-HENIN-CARVIN',
-    etiquette: 'LENS-LIEVIN-HENIN-CARVIN',
-    etiquetmin: 'Lens-Liévin-Hénin-Carvin',
-    siren_nom: 'SYNDICAT MIXTE DE MISE EN OEUVRE DU SCOT',
-    siren: '256204082',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Diagnostic',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2015-04-10',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'http://www.scot-llhc.fr/Revision',
-    siteweb2: '',
-    dialogue_n: '3',
-    dialogue_t: 'Artois - Artois Douaisis',
-    sdci: '',
-    nbre_com: '50',
-    pop: '371662',
-    surf_km2: '352',
-    info_compl:
-      "Infos au 26/11/2020 : diag en cours, le PADD sera élaboré durant l'année 2021, le DOO sera élaboré durant l'année 2022, arrêt prévu en oct 2023",
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '255902330_SCOT_00000000',
-    nom: 'SCOT DU CAMBRESIS',
-    etiquette: 'CAMBRESIS',
-    etiquetmin: 'Cambrésis',
-    siren_nom: 'SYNDICAT MIXTE DU PAYS DU CAMBRESIS',
-    siren: '255902330',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Délibération de révision du SCOT',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2018-07-11',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'http://www.paysducambresis.fr/le-pays-en-actions/urbanisme/schema-de-coherence-territoriale-scot/',
-    siteweb2: '',
-    dialogue_n: '4',
-    dialogue_t: 'Hainaut Cambrésis',
-    sdci: '',
-    nbre_com: '116',
-    pop: '165494',
-    surf_km2: '903',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: "Pôle d'équilibre territorial et rural (PETR)",
-    idurba: '200048221_SCOT_00000000',
-    nom: 'SCOT DU PAYS MARITIME ET RURAL DU MONTREUILLOIS',
-    etiquette: 'PAYS MARITIME ET RURAL DU MONTREUILLOIS',
-    etiquetmin: 'Pays maritime et rural du Montreuillois',
-    siren_nom:
-      'SYNDICAT MIXTE DU SCHEMA DE COHERENCE TERRITORIALE DU PAYS MARITIME ET RURAL DU MONTREUILLOIS',
-    siren: '200048221',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Délibération de révision du SCOT',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2020-01-06',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'https://www.ca2bm.fr/urbanisme/le-scot-schema-de-coherence-territoriale',
-    siteweb2: '',
-    dialogue_n: '5',
-    dialogue_t: 'Littoral Sud',
-    sdci: 'oui',
-    nbre_com: '95',
-    pop: '85105',
-    surf_km2: '826',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte ouvert',
-    idurba: '200039311_SCOT_00000000',
-    nom: 'SCOT BAIE DE SOMME TROIS VALLEES',
-    etiquette: 'BAIE DE SOMME TROIS VALLEES',
-    etiquetmin: 'Baie de Somme trois vallées',
-    siren_nom: 'SYNDICAT MIXTE BAIE DE SOMME 3 VALLEES',
-    siren: '200039311',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en élaboration',
-    etat_proc: "Projet d'Aménagement et de Développement Durable",
-    typeproc: '01',
-    typeproc_t: 'Elaboration',
-    typeproc_d: '2015-12-14',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'https://www.baiedesomme3vallees.fr/accueil/scot/',
-    siteweb2: '',
-    dialogue_n: '5',
-    dialogue_t: 'Littoral Sud',
-    sdci: '',
-    nbre_com: '139',
-    pop: '108593',
-    surf_km2: '1371',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: "Communauté d'agglomération",
-    idurba: '200072460_SCOT_00000000',
-    nom: "SCOT DE L'ARTOIS",
-    etiquette: 'ARTOIS',
-    etiquetmin: 'Artois',
-    siren_nom: 'CA DE BETHUNE-BRUAY, ARTOIS-LYS ROMANE',
-    siren: '200072460',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: 'Diagnostic',
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2017-09-27',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'https://www.scotdelartois.fr/revision-scot-artois/deliberation/',
-    siteweb2: '',
-    dialogue_n: '3',
-    dialogue_t: 'Artois - Artois Douaisis',
-    sdci: 'oui',
-    nbre_com: '100',
-    pop: '281249',
-    surf_km2: '647',
-    info_compl: 'Diagnostic en cours (déc 2020)',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: "Pôle d'équilibre territorial et rural (PETR)",
-    idurba: '200051118_SCOT_00000000',
-    nom: 'SCOT DU PAYS DE THIERACHE',
-    etiquette: 'PAYS DE THIERACHE',
-    etiquetmin: 'Pays de Thiérache',
-    siren_nom: 'PETR - SYNDICAT MIXTE DU PAYS DE THIERACHE',
-    siren: '200051118',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en pré-élaboration',
-    etat_proc: "SCoT en pré-élaboration (pas de délibération d'élaboration)",
-    typeproc: '',
-    typeproc_t: '',
-    typeproc_d: '',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: '',
-    siteweb2: '',
-    dialogue_n: '8',
-    dialogue_t: 'Aisne Nord - Est Somme',
-    sdci: '',
-    nbre_com: '160',
-    pop: '74946',
-    surf_km2: '1669',
-    info_compl: '',
-    modif_date: '2021-02-23',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '200003333_SCOT_00000000',
-    nom: 'SCOT DU PAYS DU CHAUNOIS',
-    etiquette: 'PAYS DU CHAUNOIS',
-    etiquetmin: 'Pays du Chaunois',
-    siren_nom: 'SYNDICAT MIXTE DU PAYS CHAUNOIS',
-    siren: '200003333',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: "Projet d'Aménagement et de Développement Durable",
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2018-01-30',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'http://www.payschaunois.fr/scot/decouverte-document-orientation-objectifs-2019/',
-    siteweb2:
-      'https://abonne.lunion.fr/id75395/article/2019-06-25/des-elus-aux-abonnes-absents-au-syndicat-mixte-du-pays-chaunois',
-    dialogue_n: '8',
-    dialogue_t: 'Aisne Nord - Est Somme',
-    sdci: 'oui',
-    nbre_com: '84',
-    pop: '72361',
-    surf_km2: '678',
-    info_compl: 'Projet arrêté non abouti. Reprise début 2021 depuis le PADD',
-    modif_date: '2021-05-04',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: "Pôle d'équilibre territorial et rural (PETR)",
-    idurba: '200075679_SCOT_00000000',
-    nom: 'SCOT TERNOIS 7 VALLEES',
-    etiquette: 'TERNOIS 7 VALLEES',
-    etiquetmin: 'Ternois 7 vallées',
-    siren_nom: "POLE D'EQUILIBRE TERRITORIAL ET RURAL (PETR) TERNOIS-7 VALLEES",
-    siren: '200075679',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en élaboration',
-    etat_proc: "Délibération d'élaboration du SCOT",
-    typeproc: '01',
-    typeproc_t: 'Elaboration',
-    typeproc_d: '2021-06-24',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb: 'https://petrternois7vallees.fr/',
-    siteweb2: '',
-    dialogue_n: '5',
-    dialogue_t: 'Littoral Sud',
-    sdci: '',
-    nbre_com: '172',
-    pop: '69003',
-    surf_km2: '1138',
-    info_compl: '',
-    modif_date: '2021-09-02',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '200093086_SCOT_00000000',
-    nom: 'SCOT DU BEAUVAISIS - CLERMONTOIS',
-    etiquette: 'BEAUVAISIS - CLERMONTOIS',
-    etiquetmin: 'Beauvaisis - Clermontois',
-    siren_nom: 'SYNDICAT MIXTE DU SCOT BEAUVAISIS - CLERMONTOIS',
-    siren: '200093086',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en élaboration',
-    etat_proc: "Délibération d'élaboration du SCOT",
-    typeproc: '01',
-    typeproc_t: 'Elaboration',
-    typeproc_d: '2021-12-15',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'https://premium.courrier-picard.fr/id45951/article/2019-10-28/le-clermontois-sassocie-au-beauvaisis-pour-son-document-durbanisme',
-    siteweb2: 'http://www.pays-clermontois.fr/medias/fichier-CCC-7809.pdf',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: '',
-    nbre_com: '71',
-    pop: '141502',
-    surf_km2: '699',
-    info_compl:
-      "Délibération d'élaboration 2021-07 du 2021-12-15. Suppression de la commune d'Ansacq du périmètre suite à son intégration à la CC de la Thelloise au 01/01/2022",
-    modif_date: '2022-08-30',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Syndicat mixte fermé',
-    idurba: '200081255_SCOT_00000000',
-    nom: "SCOT DE L'OISE ET DU PLATEAU PICARD",
-    etiquette: 'OISE PLATEAU PICARD',
-    etiquetmin: 'Oise Plateau Picard',
-    siren_nom: "SYNDICAT MIXTE DE L'OISE PLATEAU PICARD",
-    siren: '200081255',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en élaboration',
-    etat_proc: "Délibération d'élaboration du SCOT",
-    typeproc: '01',
-    typeproc_t: 'Elaboration',
-    typeproc_d: '2021-11-24',
-    diag_d: '',
-    padd_d: '',
-    doo_d: '',
-    arret_d: '',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'https://www.oise.gouv.fr/Politiques-publiques/Amenagement-durable-du-territoire/La-connaissance-de-l-Oise/Porter-a-Connaissance/Donnees-locales/Schemas-de-Coherence-Territoriale-SCoT/SCoT-de-l-Oise-et-du-Plateau-Picard',
-    siteweb2: '',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: '',
-    nbre_com: '104',
-    pop: '52829',
-    surf_km2: '839',
-    info_compl:
-      "Délibération d'élaboration du 2019-04-02 remplacée par celle du 2021-11-24 (21C-02-01)",
-    modif_date: '2022-04-05',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-  {
-    id: 'Communauté de communes',
-    idurba: '200067973_SCOT_00000000',
-    nom: 'SCOT DE LA THELLOISE',
-    etiquette: 'THELLOISE',
-    etiquetmin: 'Thelloise',
-    siren_nom: 'CC THELLOISE',
-    siren: '200067973',
-    siren_nom2: '',
-    siren2: '',
-    nat_jurid2: '',
-    etat: '',
-    etat_txt: '',
-    anneeappro: '',
-    datappro: '',
-    datefin: '',
-    proc_cours: 'SCoT en révision',
-    etat_proc: "Document d'Orientation et d'Objectifs",
-    typeproc: '02',
-    typeproc_t: 'Révision',
-    typeproc_d: '2017-12-11',
-    diag_d: '2018-04-19',
-    padd_d: '',
-    doo_d: '2020-12-01',
-    arret_d: '2021-04-25',
-    delibreg: '',
-    delibreg_d: '',
-    siteweb:
-      'https://www.thelloise.fr/amenagement-du-territoire-urbanisme/scot',
-    siteweb2: '',
-    dialogue_n: '6',
-    dialogue_t: 'Oise',
-    sdci: 'oui',
-    nbre_com: '41',
-    pop: '61126',
-    surf_km2: '307',
-    info_compl:
-      'Intégration d’Ansacq au périmètre de la CC de la Thelloise au 01/01/2022 pour reprise de la procédure de révision, auparavant interrompue par délibération communautaire du 22-06-2021',
-    modif_date: '2022-08-30',
-    typedoc: 'SCOT',
-    padd: '',
-    rapport: '',
-    doo: '',
-    urlrapport: '',
-    urlpadd: '',
-    urldoo: '',
-    urlpe: '',
-    metadata:
-      'https://geocatalogue.hautsdefrance.fr/geonetwork/srv/fre/catalog.search#/metadata/ee965118-2416-4d48-b07e-bbc696f002c2',
-  },
-]
