+++ libs/feature/record/src/lib/data-view-table/data-view-table.component.html
@@ -1,8 +1,8 @@
 <div class="w-full h-full flex flex-col">
   <gn-ui-dropdown-selector
-    *ngIf="dropdownChoices$ | async as choices"
+    *ngIf="dataFacade.dropdownChoices$ | async as choices"
     [title]="'table.select.data' | translate"
-    class="mb-7 w-auto ml-auto"
+    class="mb-7 w-auto ml-auto m-1"
     extraClass="!text-primary font-sans font-medium"
     [choices]="choices"
     [showTitle]="false"
@@ -10,24 +10,24 @@
   ></gn-ui-dropdown-selector>
   <div class="relative h-full">
     <gn-ui-table
-      *ngIf="tableData$ | async"
+      *ngIf="dataFacade.tableData$ | async"
       class="overflow-auto grow"
-      [data]="tableData$ | async"
-      [activeId]="selectionId"
+      [data]="dataFacade.tableData$ | async"
+      [activeId]="dataFacade.selectionId"
       (selected)="onTableSelect($event)"
     ></gn-ui-table>
     <gn-ui-loading-mask
-      *ngIf="loading"
+      *ngIf="dataFacade.loading"
       class="absolute inset-0"
       [message]="'table.loading.data' | translate"
     ></gn-ui-loading-mask>
     <gn-ui-popup-alert
-      *ngIf="error"
+      *ngIf="dataFacade.error"
       type="warning"
       icon="error_outline"
       class="absolute m-2 inset-0"
     >
-      <span translate>{{ error }}</span>
+      <span translate>{{ dataFacade.error }}</span>
     </gn-ui-popup-alert>
   </div>
 </div>
