+++ libs/feature/record/src/lib/state/data.facade.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing'
+
+import { DataFacade } from './data.facade'
+
+describe('DataFacade', () => {
+  let service: DataFacade
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({})
+    service = TestBed.inject(DataFacade)
+  })
+
+  it('should be created', () => {
+    expect(service).toBeTruthy()
+  })
+})
